+++ Assets/Plugins/Mediabox/Source/GameManager/Editor/Utility/EnumerableUtility.cs
@@ -2,9 +2,10 @@ using System.Collections.Generic;
 
 namespace Mediabox.GameManager.Editor.Utility {
 	public static class EnumerableUtility {
-		public static void ForEach<T>(this IEnumerable<T> enumerable, System.Action<T> method) {
+		public static IEnumerable<T> ForEach<T>(this IEnumerable<T> enumerable, System.Action<T> method) {
 			foreach (var item in enumerable)
 				method(item);
+			return enumerable;
 		}
 	}
 }